Implented a new counter display
New counter implemented to make the counter appear brighter to user. Decided against the design implementation of a segment display. Currently it does not bring enough value to the project due to the time needed for implementation into project. This decision will be revisited at a later date. Added a google font Orbitron which is similar to a segment display.